refactor originating address lookup
