+++ tools/tower/src/core/backlog.rs
@@ -83,8 +83,10 @@ pub async fn process_backlog(config: &AppCfg) -> anyhow::Result<()> {
 
 pub async fn submit_or_delete(config: &AppCfg, block: VDFProof, path: PathBuf) -> Result<()> {
     // TODO: allow user to set a profile
+    dbg!("get sender");
     let mut sender = Sender::from_app_cfg(config, None).await?;
 
+    dbg!("try commit proof");
     sender.commit_proof(block.clone()).await?;
 
     match sender.eval_response() {
