+++ types/src/type_extensions/client_ext.rs
@@ -1,6 +1,7 @@
 use crate::legacy_types::app_cfg::AppCfg;
 use crate::type_extensions::cli_config_ext::CliConfigExt;
 use crate::util::parse_function_id;
+use crate::exports::AuthenticationKey;
 
 use anyhow::{anyhow, Context};
 use async_trait::async_trait;
@@ -46,6 +47,9 @@ pub trait ClientExt {
     async fn find_good_upstream(list: Vec<Url>) -> anyhow::Result<(Client, ChainId)>;
 
     fn from_vendor_config() -> anyhow::Result<Client>;
+
+    async fn lookup_originating_address(&self, authentication_key: AuthenticationKey,
+    ) -> anyhow::Result<AccountAddress>;
 
     async fn get_move_resource<T: MoveStructType + DeserializeOwned>(
         &self,
@@ -122,6 +126,32 @@ impl ClientExt for Client {
         Ok(Client::new(Url::from_str(&rest_url).unwrap()))
     }
 
+    /// Addresses will diverge from the keypair which originally created the address.
+    /// The Address and AuthenticationKey key are the same bytes: a sha3 hash
+    /// of the public key. If you rotate the keypair for that address, the implied (derived) public key, and thus authentication key will not be the same as the
+    ///  Origial/originating address. For this reason, we need to look up the original address
+    /// Addresses are stored in the OriginatingAddress table, which is a table
+    /// that maps a derived address to the original address. This function
+    /// looks up the original address for a given derived address.
+    async fn lookup_originating_address(
+        &self,
+        authentication_key: AuthenticationKey,
+    ) -> anyhow::Result<AccountAddress> {
+        // the move View will return the same address_key if it has an unmodified Authkey (never been rotated)
+        // let bytes = authentication_key.to_vec();
+        // let cast_address = AccountAddress::from_bytes(bytes.as_slice())?;
+
+        let function_id = entry_function_id("account", "get_originating_address")?;
+        let request = ViewRequest {
+            function: function_id,
+            type_arguments: vec![],
+            arguments: vec![authentication_key.to_string().into()],
+        };
+
+        let res = self.view(&request, None).await?.into_inner();
+        let addr = serde_json::from_value(res[0].clone())?;
+        Ok(addr)
+    }
     // async fn get_account_balance_libra(&self, account: AccountAddress) -> Result<SlowWalletBalance> {
 
     //   let slow_balance_id = entry_function_id("slow_wallet", "balance")?;
